Verify vTPM creation after svc restart
Ensure that vTPM creation is still possible after restarting the nova
libvirt service.

Also added a new client service for virtqemud that can start/stop the
service. Updated the nova waiter to allow for checking status for either
nova_compute or virtqemud. Originally the nova waiter would rely on just
looking at results of compute services, specifically the state field to
determine if nova_compute is up or down. Extend the nova waiter to check
either the state field or the status field based on the service that is
calling.

Change-Id: I5895602b33c78698beec363fbc00bca876b3cefa