Update build exception when using unified limits
When Unified Limits are enabled, the invalid vtpm models being tested
downstream in test_invalid_model_version_creation is now throwing a new
exception. This is because Unified Limits will call from_request_spec
and confirm the flavor extra_specs are valid earlier in the path. The
resulting error now throws a ServerFault from Nova-Api.  Updating the
test so that when Unified Limits are configured, it will check for
ServerFault when trying to build with invalid flavor specs, but still
check for BuildExcpetion if Unified Limits have not been enabled.

Change-Id: I412f55327d475ee5bc7a9a34c9e5012d1de3c3e0