Disable uefi boot support in gate
Currently provided flavor of guests from nodepool do not appear to have
the same support for NVRAM as it once did. This is causing UEFI related
tests to fail. Disabling until we can root cause issue.

Also documented an upstream bug [1] to help confirm if this is an issue
with nodepool or an issue with a change in Nova.

[1] https://bugs.launchpad.net/nova/+bug/2137560

Change-Id: I22545341a5e706f53235eb58db6cde03abfb76a5
Signed-off-by: jamepark4 <jparker@redhat.com>